Fix device testing bugs: onboarding flow, session persistence, scroll conflicts, empty states, and data type mismatches
- Fix onboarding screen showing after biometric unlock for authenticated users
- Reorder biometric check before app initialization in DriftApp
- Fix isLoading initial state to prevent UI flash on launch
- Persist user profile to keychain for session restoration
- Add user profile endpoint and fetch fallback in AuthService
- Fix scroll conflicts: replace DragGesture(minimumDistance: 0) with Button/onTapGesture in charts and leaky bucket cards
- Add Face ID usage description to Info.plist
- Fix spending empty state to hide hero card and chart when no transactions
- Fix leaky bucket detector returning mock data for empty real transactions
- Make leaky bucket header dynamic based on bucket count
- Fix Double vs Decimal type mismatch in backend summary DTOs
- Add styled profile section with initials to Settings view

+++ DriftBackend/Sources/App/Models/Transaction.swift
@@ -80,7 +80,7 @@ extension Transaction {
             id: id!,
             accountId: $account.id,
             plaidTransactionId: plaidTransactionId,
-            amount: amount,
+            amount: Decimal(amount),
             date: date,
             merchantName: merchantName,
             category: category,
@@ -95,7 +95,7 @@ struct TransactionDTO: Content {
     let id: UUID
     let accountId: UUID
     let plaidTransactionId: String?
-    let amount: Double
+    let amount: Decimal
     let date: Date
     let merchantName: String
     let category: String
